chore: 消除项目waring
Description: 修改对应的数据类型，消除waring

Log: 消除项目waring

link:

+++ src/scheduleview.h
@@ -82,7 +82,7 @@ private:
     int                       m_leftMagin;
     int                       m_topMagin;
     int                       m_rightmagin = 0;
-    int                       m_TotalDay;
+    qint64                       m_TotalDay;
     QDate                     m_currteDate;
     int                       m_firstWeekDay;
     QDate                     m_beginDate;
